+++ RPG/BattleWithRetreat.cpp
@@ -58,12 +58,12 @@ void BattleWithRetreat::fight(Character^ character)
 		switch (actionAnswer)
 		{
 		case 0:		//Melee attack
-			attackPower = attackModifier * character->baseMeleeAttack() - opponentBlockPower;
+			attackPower = safe_cast<int>(attackModifier * character->baseMeleeAttack() - opponentBlockPower);
 			chosenOpponent->getStatistics()->substract(StatisticsFactory::onlyHp(attackPower));
 
 			if (chosenOpponent->getStatistics()->HP > 0)
 			{
-				attackPower = attackModifier * chosenOpponent->baseMeleeAttack() - playerBlockPower;
+				attackPower = safe_cast<int>(attackModifier * chosenOpponent->baseMeleeAttack() - playerBlockPower);
 				character->getStatistics()->substract(StatisticsFactory::onlyHp(attackPower));
 			}
 
@@ -71,12 +71,12 @@ void BattleWithRetreat::fight(Character^ character)
 			opponentBlockPower = 0;
 			break;
 		case 1:		//Ranged attack
-			attackPower = attackModifier * character->baseRangeAttack();
+			attackPower = safe_cast<int>(attackModifier * character->baseRangeAttack());
 			chosenOpponent->getStatistics()->substract(StatisticsFactory::onlyHp(attackPower));
 
 			if (chosenOpponent->getStatistics()->HP > 0)
 			{
-				attackPower = attackModifier * chosenOpponent->baseRangeAttack();
+				attackPower = safe_cast<int>(attackModifier * chosenOpponent->baseRangeAttack());
 				character->getStatistics()->substract(StatisticsFactory::onlyHp(attackPower));
 			}
